fix cyclic dependency
