+++ extensions/common/vault/vault-filesystem/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -13,3 +13,4 @@
 #
 
 org.eclipse.edc.vault.filesystem.FsVaultExtension
+org.eclipse.edc.vault.filesystem.JskPrivateKeyResolverExtension
